Many things were tested (a few bugs were removed) Modification: the first EmotionAwareness Message now occurs after the module has started which allow to put the module in fullscreen if needed Addition: A Message is sent when the module is started (usefull to know when exactly the participants have started the collaboration)
git-svn-id: https://subversion.unige.ch/users/chanel/DrewLite@17 7b2cfa2a-5ca0-44b3-9680-3171f8a568b4

+++ Drew/Client/MultiModules/TwoModules.java
@@ -119,7 +119,7 @@ public class TwoModules extends Panel implements CooperativeModule {
                     if (ratioModA >= 1)
                         ratioModA = 0.5;
                 }
-                catch(Exception e) //null string, or not integer format
+                catch(Exception e) //null string, or not float format
                 {
                     ratioModA = 0.5;
                 }
@@ -163,8 +163,8 @@ public class TwoModules extends Panel implements CooperativeModule {
 		}
 		add((Panel)modA); modA.init();
 		add((Panel)modB); modB.init();
-	}
-  
+        }
+
 	/**
         * modules are started
         */
